refactor(auth-mapping): rename authmappingutils file
