+++ src/components/tables/ProjectTable.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Button from '@gen3/ui-component/dist/components/Button';
 import Table from './base/Table';
 import { useArboristUI } from '../../configs';
-import { userHasMethodOnProject } from '../../utilsAuth';
+import { userHasMethodOnProject } from '../../authMappingUtils';
 import './ProjectTable.less';
 
 function compare(a, b) {
