+++ src/components/tables/TransactionLogTable.jsx
@@ -5,7 +5,7 @@ import Spinner from '../Spinner';
 import { humanFileSize } from '../../utils.js';
 import './TransactionLogTable.less';
 import { useArboristUI } from '../../configs';
-import { userHasMethodOnAnyProject } from '../../utilsAuth';
+import { userHasMethodOnAnyProject } from '../../authMappingUtils';
 
 const formatText = text => text[0] + text.slice(1).toLowerCase();
